feat: 친구(ARCH-12)

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/data/dto/friend/request/FriendAcceptRequestDto.kt
@@ -1,5 +1,5 @@
 package com.droidblossom.archive.data.dto.friend.request
 
 data class FriendAcceptRequestDto (
-    val friendId : Double
+    val friendId : Long
 )
