+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/domain/model/friend/FriendReqRequest.kt
@@ -3,7 +3,7 @@ package com.droidblossom.archive.domain.model.friend
 import com.droidblossom.archive.data.dto.friend.request.FriendReqRequestDto
 
 data class FriendReqRequest(
-    val friendId : Double
+    val friendId : Long
 ){
     fun toDto() = FriendReqRequestDto(
         friendId = this.friendId,
